code/customer-web-server/requirements.txt
# Flask web framework
flask==3.0.0

# Kafka client for Python (kafka-python-ng is the maintained fork for Python 3.12+)
kafka-python-ng==2.2.2

# HTTP client for calling Customer Management API
requests==2.31.0
